+++ setup.cfg
@@ -129,8 +129,8 @@ exclude =
 max-line-length = 120
 exclude =
     venv/*
-    src/helm/proxy/clients/dalle_mini/*
-    src/helm/proxy/clients/mindalle/*
+    src/helm/proxy/clients/image_generation/dalle_mini/*
+    src/helm/proxy/clients/image_generation/mindalle/*
 
 # Ignore completely:
 # E203 - White space before ':', (conflicts with black)
