+++ src/helm/proxy/clients/image_generation/adobe_vision_client.py
@@ -9,44 +9,27 @@ from helm.common.tokenization_request import (
     DecodeRequestResult,
 )
 
-from .client import Client
+from helm.proxy.clients.client import Client
 
 
-class AlephAlphaVisionClient(Client):
+class AdobeVisionClient(Client):
     """
-    Client for Aleph Alpha vision models. Offline eval only.
+    Client for Adobe vision models. Offline eval only.
     """
 
-    DEFAULT_IMAGE_HEIGHT: int = 512
-    DEFAULT_IMAGE_WIDTH: int = 512
-
-    DEFAULT_GUIDANCE_SCALE: float = 7.5
-    DEFAULT_STEPS: int = 50
+    SUPPORTED_MODELS: List[str] = ["giga-gan", "firefly"]
 
     @staticmethod
     def convert_to_raw_request(request: Request) -> Dict:
+        # Use default hyperparameters for everything else
         raw_request: Dict = {
             "request_type": "image-model-inference",
             "model": request.model_engine,
             "prompt": request.prompt,
             "n": request.num_completions,
-            "guidance_scale": AlephAlphaVisionClient.DEFAULT_GUIDANCE_SCALE,
-            "steps": AlephAlphaVisionClient.DEFAULT_STEPS,
-            "width": AlephAlphaVisionClient.DEFAULT_IMAGE_WIDTH,
-            "height": AlephAlphaVisionClient.DEFAULT_IMAGE_HEIGHT,
         }
         if request.random is not None:
             raw_request["random"] = request.random
-
-        if isinstance(request, TextToImageRequest):
-            if request.guidance_scale is not None:
-                raw_request["guidance_scale"] = request.guidance_scale
-            if request.steps is not None:
-                raw_request["steps"] = request.steps
-            if request.width is not None and request.height is not None:
-                raw_request["width"] = request.width
-                raw_request["height"] = request.height
-
         return raw_request
 
     def __init__(self, cache_config: CacheConfig):
@@ -58,10 +41,10 @@ class AlephAlphaVisionClient(Client):
         if not isinstance(request, TextToImageRequest):
             raise ValueError(f"Wrong type of request: {request}")
 
-        if request.model_engine != "m-vader":
+        if request.model_engine not in self.SUPPORTED_MODELS:
             raise ValueError(f"Unsupported model: {request.model_engine}")
 
-        raw_request = AlephAlphaVisionClient.convert_to_raw_request(request)
+        raw_request = AdobeVisionClient.convert_to_raw_request(request)
         raw_request.pop("random", None)
         cache_key: Dict = Client.make_cache_key(raw_request, request)
 
@@ -74,7 +57,7 @@ class AlephAlphaVisionClient(Client):
 
             response, cached = self._cache.get(cache_key, fail)
         except RuntimeError as e:
-            error: str = f"AlephAlphaVisionClient error: {e}"
+            error: str = f"Adobe Vision Client error: {e}"
             return RequestResult(success=False, cached=False, error=error, completions=[], embedding=[])
 
         completions: List[Sequence] = [
