+++ src/helm/proxy/clients/image_generation/cogview2/cluster_label.npy
