+++ src/helm/proxy/clients/image_generation/cogview2/coglm_strategy.py
