+++ src/helm/proxy/clients/image_generation/cogview2/sr_pipeline/__init__.py
