+++ src/helm/proxy/clients/image_generation/cogview2/sr_pipeline/dsr_model.py
