+++ src/helm/proxy/clients/image_generation/cogview2_client.py
@@ -11,7 +11,7 @@ from helm.common.tokenization_request import (
     TokenizationRequest,
     TokenizationRequestResult,
 )
-from .client import Client, wrap_request_time
+from helm.proxy.clients.client import Client, wrap_request_time
 
 import os
 import torch
@@ -20,9 +20,13 @@ from functools import partial
 from torchvision.utils import save_image
 from SwissArmyTransformer import get_args
 from SwissArmyTransformer.generation.autoregressive_sampling import filling_sequence
-from .cogview2.coglm_strategy import CoglmStrategy
-from .cogview2.sr_pipeline import SRGroup
-from .cogview2.coglm_utils import get_masks_and_position_ids_coglm, get_recipe, InferenceModel
+from helm.proxy.clients.image_generation.cogview2.coglm_strategy import CoglmStrategy
+from helm.proxy.clients.image_generation.cogview2.sr_pipeline import SRGroup
+from helm.proxy.clients.image_generation.cogview2.coglm_utils import (
+    get_masks_and_position_ids_coglm,
+    get_recipe,
+    InferenceModel,
+)
 from icetk import icetk as tokenizer
 
 tokenizer.add_special_tokens(["<start_of_image>", "<start_of_english>", "<start_of_chinese>"])
