+++ src/helm/proxy/clients/image_generation/dalle_mini/__init__.py
