+++ src/helm/proxy/clients/image_generation/dalle_mini/model/__init__.py
