+++ src/helm/proxy/clients/image_generation/dalle_mini/model/tokenizer.py
