+++ src/helm/proxy/clients/image_generation/dalle_mini/vqgan_jax/__init__.py
