+++ src/helm/proxy/clients/image_generation/dalle_mini/vqgan_jax/configuration_vqgan.py
