+++ src/helm/proxy/clients/image_generation/dalle_mini/vqgan_jax/convert_pt_model_to_jax.py
