+++ src/helm/proxy/clients/image_generation/dalle_mini/vqgan_jax/modeling_flax_vqgan.py
