+++ src/helm/proxy/clients/image_generation/mindalle/models/__init__.py
