+++ src/helm/proxy/clients/image_generation/mindalle/models/stage1/layers.py
