+++ src/helm/proxy/clients/image_generation/mindalle/models/stage2/transformer.py
