+++ src/helm/proxy/clients/image_generation/mindalle/models/tokenizer.py
