+++ src/helm/proxy/clients/image_generation/mindalle/utils/__init__.py
