+++ src/helm/proxy/clients/image_generation/mindalle/utils/config.py
