+++ src/helm/proxy/clients/image_generation/nudity_check_client.py
